factor out collect-algorithm-diagnostics
